Extract side-effect checks into ExprSideEffectsHelper
Move expression side-effect detection logic from TypeSpecifier into a
dedicated ExprSideEffectsHelper class with public rememberFuncCall,
rememberMethodCall, rememberStaticCall, and subExpressionsHaveSideEffects
methods that encapsulate all purity and side-effect conditions.
